refactor(sea): extract runtime node handlers
