+++ sea/runtime.py
@@ -24,6 +24,13 @@ from sea.runtime_context import preview_context as preview_context_impl
 from sea.runtime_graph import compile_with_langgraph as compile_with_langgraph_impl
 from sea.runtime_llm import lg_llm_node as lg_llm_node_impl
 from sea.runtime_runner import run_playbook
+from sea.runtime_nodes import (
+    lg_exec_node as lg_exec_node_impl,
+    lg_stelis_end_node as lg_stelis_end_node_impl,
+    lg_stelis_start_node as lg_stelis_start_node_impl,
+    lg_subplay_node as lg_subplay_node_impl,
+    lg_tool_call_node as lg_tool_call_node_impl,
+)
 from sea.runtime_state import (
     apply_output_mapping,
     eval_arithmetic_expression,
@@ -518,107 +525,7 @@ class SEARuntime:
         return self._runtime_engine.lg_tool_node(node_def, persona, playbook, event_callback)
 
     def _lg_tool_call_node(self, node_def: Any, persona: Any, playbook: PlaybookSchema, event_callback: Optional[Callable[[Dict[str, Any]], None]] = None):
-        """Execute a tool dynamically based on an LLM node's tool call decision.
-
-        Reads tool name and arguments from state (stored by an LLM node with
-        available_tools), looks up the tool in TOOL_REGISTRY, and executes it.
-        This enables agentic loops without per-tool branching.
-        """
-        from tools import TOOL_REGISTRY
-        from tools.context import persona_context
-
-        call_source = getattr(node_def, "call_source", "fc") or "fc"
-        output_key = getattr(node_def, "output_key", None)
-
-        async def node(state: dict):
-            # Check for cancellation
-            cancellation_token = state.get("_cancellation_token")
-            if cancellation_token:
-                cancellation_token.raise_if_cancelled()
-
-            node_id = getattr(node_def, "id", "tool_call")
-            if event_callback:
-                event_callback({"type": "status", "content": f"{playbook.name} / {node_id}", "playbook": playbook.name, "node": node_id})
-
-            # Resolve tool name and args from state
-            tool_name = self._resolve_state_value(state, f"{call_source}.name")
-            tool_args = self._resolve_state_value(state, f"{call_source}.args")
-
-            # Fallback to legacy state keys
-            if not tool_name:
-                tool_name = state.get("tool_name", "")
-                tool_args = state.get("tool_args", {})
-
-            if not tool_name:
-                error_msg = f"[sea][tool_call] No tool name found in state (call_source={call_source})"
-                LOGGER.error(error_msg)
-                state["last"] = error_msg
-                if output_key:
-                    state[output_key] = error_msg
-                return state
-
-            if not isinstance(tool_args, dict):
-                LOGGER.warning("[sea][tool_call] tool_args is not a dict (%s), using empty args", type(tool_args).__name__)
-                tool_args = {}
-
-            tool_func = TOOL_REGISTRY.get(tool_name)
-            if tool_func is None:
-                error_msg = f"[sea][tool_call] Tool '{tool_name}' not found in registry"
-                LOGGER.error(error_msg)
-                state["last"] = error_msg
-                if output_key:
-                    state[output_key] = error_msg
-                return state
-
-            persona_obj = state.get("persona_obj") or persona
-            persona_id = getattr(persona_obj, "persona_id", "unknown")
-
-            try:
-                persona_dir = getattr(persona_obj, "persona_log_path", None)
-                persona_dir = persona_dir.parent if persona_dir else Path.cwd()
-                manager_ref = getattr(persona_obj, "manager_ref", None)
-
-                LOGGER.info("[sea][tool_call] CALL %s (persona=%s) args=%s", tool_name, persona_id, tool_args)
-
-                if persona_id and persona_dir:
-                    with persona_context(persona_id, persona_dir, manager_ref, playbook_name=playbook.name):
-                        result = tool_func(**tool_args)
-                else:
-                    result = tool_func(**tool_args)
-
-                result_str = str(result)
-                result_preview = result_str[:500] + "..." if len(result_str) > 500 else result_str
-                LOGGER.info("[sea][tool_call] RESULT %s -> %s", tool_name, result_preview)
-                log_sea_trace(playbook.name, node_id, "TOOL_CALL", f"action={tool_name} args={tool_args} → {result_str}")
-
-                # Activity trace
-                if not playbook.name.startswith(("meta_", "sub_")):
-                    pb_display = playbook.display_name or playbook.name
-                    _at = state.get("_activity_trace")
-                    if isinstance(_at, list):
-                        _at.append({"action": "tool_call", "name": tool_name, "playbook": pb_display})
-                    if event_callback:
-                        event_callback({
-                            "type": "activity", "action": "tool_call", "name": tool_name,
-                            "playbook": pb_display, "status": "completed",
-                            "persona_id": getattr(persona, "persona_id", None),
-                            "persona_name": getattr(persona, "persona_name", None),
-                        })
-
-                state["last"] = result_str
-                if output_key:
-                    state[output_key] = result
-
-            except Exception as exc:
-                error_msg = f"Tool error ({tool_name}): {exc}"
-                state["last"] = error_msg
-                if output_key:
-                    state[output_key] = error_msg
-                LOGGER.exception("[sea][tool_call] %s failed", tool_name)
-
-            return state
-
-        return node
+        return lg_tool_call_node_impl(self, node_def, persona, playbook, event_callback)
 
     def _lg_exec_node(
         self,
@@ -630,9 +537,7 @@ class SEARuntime:
         outputs: Optional[List[str]] = None,
         event_callback: Optional[Callable[[Dict[str, Any]], None]] = None,
     ):
-        return self._runtime_engine.lg_exec_node(
-            node_def, playbook, persona, building_id, auto_mode, outputs, event_callback
-        )
+        return lg_exec_node_impl(self, node_def, playbook, persona, building_id, auto_mode, outputs, event_callback)
 
     def _lg_memorize_node(self, node_def: Any, persona: Any, playbook: PlaybookSchema, outputs: Optional[List[str]] = None, event_callback: Optional[Callable[[Dict[str, Any]], None]] = None):
         return self._runtime_engine.lg_memorize_node(node_def, persona, playbook, outputs, event_callback)
@@ -705,96 +610,7 @@ class SEARuntime:
         return state
 
     def _lg_subplay_node(self, node_def: Any, persona: Any, building_id: str, playbook: PlaybookSchema, auto_mode: bool, outputs: Optional[List[str]] = None, event_callback: Optional[Callable[[Dict[str, Any]], None]] = None):
-        async def node(state: dict):
-            # Check for cancellation at start of node
-            cancellation_token = state.get("_cancellation_token")
-            if cancellation_token:
-                cancellation_token.raise_if_cancelled()
-            
-            # Send status event for node execution
-            node_id = getattr(node_def, "id", "subplay")
-            if event_callback:
-                event_callback({"type": "status", "content": f"{playbook.name} / {node_id}", "playbook": playbook.name, "node": node_id})
-            # Get subplaybook name
-            sub_name = getattr(node_def, "playbook", None) or getattr(node_def, "action", None)
-            if not sub_name:
-                msg = "(sub-playbook missing name)"
-                state["last"] = msg
-                return state
-
-            # Load subplaybook
-            sub_pb = self._load_playbook_for(sub_name, persona, building_id)
-            if not sub_pb:
-                msg = f"Sub-playbook {sub_name} not found"
-                state["last"] = msg
-                return state
-
-            # Format input template with state variables
-            template = getattr(node_def, "input_template", "{input}") or "{input}"
-            variables = dict(state)
-            variables.update({
-                "input": state.get("inputs", {}).get("input", ""),
-                "last": state.get("last", ""),
-            })
-            from .runtime_utils import _format as runtime_format
-            sub_input = runtime_format(template, variables)
-            eff_bid = self._effective_building_id(persona, building_id)
-
-            # Determine execution mode
-            execution = getattr(node_def, "execution", "inline") or "inline"
-            subagent_thread_id = None
-            subagent_parent_id = None
-
-            if execution == "subagent":
-                label = f"Subagent: {sub_name}"
-                subagent_thread_id, subagent_parent_id = self._start_subagent_thread(persona, label=label)
-                if not subagent_thread_id:
-                    LOGGER.warning("[sea][subplay] Failed to start subagent thread for '%s', falling back to inline", sub_name)
-                    execution = "inline"  # Fallback
-                else:
-                    log_sea_trace(playbook.name, node_id, "SUBPLAY", f"→ {sub_name} [subagent thread={subagent_thread_id}] (input=\"{str(sub_input)}\")")
-
-            if execution == "inline":
-                log_sea_trace(playbook.name, node_id, "SUBPLAY", f"→ {sub_name} (input=\"{str(sub_input)}\")")
-
-            # Execute subplaybook
-            # Note: We call _run_playbook directly (not via asyncio.to_thread) to keep
-            # SQLite connections on the same thread. _run_playbook handles its own
-            # async/sync boundary internally via ThreadPoolExecutor.
-            try:
-                sub_outputs = self._run_playbook(sub_pb, persona, eff_bid, sub_input, auto_mode, True, state, event_callback)
-            except LLMError:
-                LOGGER.exception("[sea][subplay] LLM error in subplaybook '%s'", sub_name)
-                if execution == "subagent" and subagent_thread_id:
-                    self._end_subagent_thread(persona, subagent_thread_id, subagent_parent_id, generate_chronicle=False)
-                raise
-            except Exception as exc:
-                LOGGER.exception("[sea][subplay] Failed to execute subplaybook '%s'", sub_name)
-                # End subagent thread on error (no chronicle)
-                if execution == "subagent" and subagent_thread_id:
-                    self._end_subagent_thread(persona, subagent_thread_id, subagent_parent_id, generate_chronicle=False)
-                state["last"] = f"Sub-playbook error: {exc}"
-                return state
-
-            # End subagent thread on success
-            if execution == "subagent" and subagent_thread_id:
-                gen_chronicle = getattr(node_def, "subagent_chronicle", True)
-                chronicle = self._end_subagent_thread(persona, subagent_thread_id, subagent_parent_id, generate_chronicle=gen_chronicle)
-                state["_subagent_chronicle"] = chronicle or ""
-                log_sea_trace(playbook.name, node_id, "SUBPLAY", f"← {sub_name} [subagent ended, chronicle={'yes' if chronicle else 'no'}]")
-
-            last_text = sub_outputs[-1] if sub_outputs else ""
-            state["last"] = last_text
-
-            # Propagate outputs if requested
-            if getattr(node_def, "propagate_output", False) and sub_outputs and outputs is not None:
-                outputs.extend(sub_outputs)
-
-            # Note: State variables are propagated via output_schema in _compile_with_langgraph
-            # No special handling needed here anymore
-
-            return state
-        return node
+        return lg_subplay_node_impl(self, node_def, persona, building_id, playbook, auto_mode, outputs, event_callback)
 
     def _lg_set_node(self, node_def: Any, playbook: PlaybookSchema, event_callback: Optional[Callable[[Dict[str, Any]], None]] = None):
         """Create a node that sets/modifies state variables."""
@@ -929,134 +745,7 @@ class SEARuntime:
         playbook: PlaybookSchema,
         event_callback: Optional[Callable[[Dict[str, Any]], None]] = None
     ):
-        """Create a node that starts a new Stelis thread for hierarchical context management."""
-
-        async def node(state: dict):
-            # Check for cancellation
-            cancellation_token = state.get("_cancellation_token")
-            if cancellation_token:
-                cancellation_token.raise_if_cancelled()
-
-            node_id = getattr(node_def, "id", "stelis_start")
-            label_raw = getattr(node_def, "label", None) or "Stelis Session"
-            from .runtime_utils import _format as runtime_format
-            label = runtime_format(label_raw, state)
-
-            # Send status event
-            if event_callback:
-                event_callback({
-                    "type": "status",
-                    "content": f"{playbook.name} / {node_id}",
-                    "playbook": playbook.name,
-                    "node": node_id
-                })
-
-            # Get Stelis configuration
-            stelis_config = getattr(node_def, "stelis_config", None) or {}
-            if hasattr(stelis_config, "__dict__"):
-                # Convert Pydantic model to dict if needed
-                stelis_config = {
-                    "window_ratio": getattr(stelis_config, "window_ratio", 0.8),
-                    "max_depth": getattr(stelis_config, "max_depth", 3),
-                    "chronicle_prompt": getattr(stelis_config, "chronicle_prompt", None),
-                }
-
-            window_ratio = stelis_config.get("window_ratio", 0.8)
-            max_depth = stelis_config.get("max_depth", 3)
-            chronicle_prompt = stelis_config.get("chronicle_prompt")
-
-            # Get memory adapter from persona
-            memory_adapter = getattr(persona, "sai_memory", None)
-            if not memory_adapter:
-                LOGGER.warning("[stelis] No memory adapter found for persona %s", persona.persona_id)
-                state["stelis_error"] = "No memory adapter available"
-                state["stelis_available"] = False
-                return state
-
-            # Check if we can start a new Stelis thread
-            if not memory_adapter.can_start_stelis(max_depth=max_depth):
-                error_msg = f"Stelis max depth exceeded (max={max_depth})"
-                LOGGER.warning("[stelis] %s for persona %s", error_msg, persona.persona_id)
-                state["stelis_error"] = error_msg
-                state["stelis_available"] = False
-                return state
-
-            # Get current thread as parent
-            # get_current_thread() returns full thread ID (e.g., "air_city_a:__persona__")
-            # Use it directly as parent_thread_id, don't pass to _thread_id which adds prefix
-            parent_thread_id = memory_adapter.get_current_thread()
-            if parent_thread_id is None:
-                # Fallback to default persona thread if no active thread set
-                parent_thread_id = memory_adapter._thread_id(None)
-
-            # Create new Stelis thread
-            stelis = memory_adapter.start_stelis_thread(
-                parent_thread_id=parent_thread_id,
-                window_ratio=window_ratio,
-                chronicle_prompt=chronicle_prompt,
-                max_depth=max_depth,
-                label=label,
-            )
-
-            if not stelis:
-                LOGGER.error("[stelis] Failed to create Stelis thread for persona %s", persona.persona_id)
-                state["stelis_error"] = "Failed to create Stelis thread"
-                state["stelis_available"] = False
-                return state
-
-            # Add anchor message to PARENT thread (before switching)
-            # This message will be dynamically expanded when viewing parent thread
-            import time
-            anchor_message = {
-                "role": "system",
-                "content": "",  # Content is dynamically generated
-                "metadata": {
-                    "type": "stelis_anchor",
-                    "stelis_thread_id": stelis.thread_id,
-                    "stelis_label": label,
-                    "created_at": int(time.time()),
-                },
-                "embedding_chunks": 0,  # Don't embed this message
-            }
-            memory_adapter.append_persona_message(
-                anchor_message,
-                thread_suffix=parent_thread_id.split(":")[-1] if ":" in parent_thread_id else parent_thread_id,
-            )
-            LOGGER.debug(
-                "[stelis] Added anchor message to parent thread %s for Stelis %s",
-                parent_thread_id, stelis.thread_id
-            )
-
-            # Switch to new Stelis thread
-            memory_adapter.set_active_thread(stelis.thread_id)
-            log_sea_trace(playbook.name, node_id, "STELIS_START", f"thread={stelis.thread_id} label=\"{label}\"")
-
-            # Update state with Stelis info
-            state["stelis_thread_id"] = stelis.thread_id
-            state["stelis_parent_thread_id"] = parent_thread_id
-            state["stelis_depth"] = stelis.depth
-            state["stelis_window_ratio"] = window_ratio
-            state["stelis_label"] = label
-            state["stelis_available"] = True
-
-            LOGGER.info(
-                "[stelis] Started Stelis thread %s (parent=%s, depth=%d, ratio=%.2f, label=%s)",
-                stelis.thread_id, parent_thread_id, stelis.depth, window_ratio, label
-            )
-
-            # Emit event for UI
-            if event_callback:
-                event_callback({
-                    "type": "stelis_start",
-                    "thread_id": stelis.thread_id,
-                    "parent_thread_id": parent_thread_id,
-                    "depth": stelis.depth,
-                    "label": label,
-                })
-
-            return state
-
-        return node
+        return lg_stelis_start_node_impl(self, node_def, persona, playbook, event_callback)
 
     def _lg_stelis_end_node(
         self,
@@ -1065,102 +754,7 @@ class SEARuntime:
         playbook: PlaybookSchema,
         event_callback: Optional[Callable[[Dict[str, Any]], None]] = None
     ):
-        """Create a node that ends the current Stelis thread and returns to parent context."""
-
-        async def node(state: dict):
-            # Check for cancellation
-            cancellation_token = state.get("_cancellation_token")
-            if cancellation_token:
-                cancellation_token.raise_if_cancelled()
-
-            node_id = getattr(node_def, "id", "stelis_end")
-            generate_chronicle = getattr(node_def, "generate_chronicle", True)
-
-            # Send status event
-            if event_callback:
-                event_callback({
-                    "type": "status",
-                    "content": f"{playbook.name} / {node_id}",
-                    "playbook": playbook.name,
-                    "node": node_id
-                })
-
-            # Get memory adapter from persona
-            memory_adapter = getattr(persona, "sai_memory", None)
-            if not memory_adapter:
-                LOGGER.warning("[stelis] No memory adapter found for persona %s", persona.persona_id)
-                return state
-
-            # Get current Stelis thread info from state
-            current_thread_id = state.get("stelis_thread_id")
-            parent_thread_id = state.get("stelis_parent_thread_id")
-
-            if not current_thread_id or not parent_thread_id:
-                LOGGER.warning("[stelis] STELIS_END called without active Stelis context")
-                return state
-
-            # Verify we're in a Stelis thread
-            stelis_info = memory_adapter.get_stelis_info(current_thread_id)
-            if not stelis_info:
-                LOGGER.warning("[stelis] Current thread %s is not a Stelis thread", current_thread_id)
-                return state
-
-            # Generate Chronicle summary if requested
-            chronicle_summary = None
-            if generate_chronicle:
-                chronicle_summary = self._generate_stelis_chronicle(
-                    persona,
-                    current_thread_id,
-                    stelis_info.chronicle_prompt
-                )
-                LOGGER.info(
-                    "[stelis] Generated Chronicle for thread %s: %s...",
-                    current_thread_id,
-                    chronicle_summary[:100] if chronicle_summary else "(empty)"
-                )
-
-            # End the Stelis thread
-            success = memory_adapter.end_stelis_thread(
-                thread_id=current_thread_id,
-                status="completed",
-                chronicle_summary=chronicle_summary,
-            )
-
-            if not success:
-                LOGGER.error("[stelis] Failed to end Stelis thread %s", current_thread_id)
-
-            # Switch back to parent thread
-            memory_adapter.set_active_thread(parent_thread_id)
-
-            # Store Chronicle in state for potential use
-            if chronicle_summary:
-                state["stelis_chronicle"] = chronicle_summary
-
-            # Clear Stelis state
-            state["stelis_thread_id"] = None
-            state["stelis_parent_thread_id"] = None
-            state["stelis_depth"] = None
-
-            LOGGER.info(
-                "[stelis] Ended Stelis thread %s, returned to parent %s",
-                current_thread_id, parent_thread_id
-            )
-
-            _chron_str = chronicle_summary or "(none)"
-            log_sea_trace(playbook.name, node_id, "STELIS_END", f"thread={current_thread_id} chronicle=\"{_chron_str}\"")
-
-            # Emit event for UI
-            if event_callback:
-                event_callback({
-                    "type": "stelis_end",
-                    "thread_id": current_thread_id,
-                    "parent_thread_id": parent_thread_id,
-                    "chronicle_generated": generate_chronicle,
-                })
-
-            return state
-
-        return node
+        return lg_stelis_end_node_impl(self, node_def, persona, playbook, event_callback)
 
     def _generate_stelis_chronicle(
         self,
